Clean up and moved actions to the framework
