+++ ShortcutsSwift.playgroundbook/Contents/Sources/Actions/GetCurrentSong.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public func getCurrentSong() -> ActionWithOutput {
+    return Action(identifier: "is.workflow.actions.getcurrentsong", parameters: [:])
+}
