+++ ShortcutsSwift.playgroundbook/Contents/Sources/Actions/GetText.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public func getText(_ text: String) -> ActionWithOutput {
+    return Action(
+        identifier: "is.workflow.actions.gettext",
+        parameters: ["WFTextActionText": withInterpolatedText(text)]
+    )
+}
