+++ ShortcutsSwift.playgroundbook/Contents/Sources/Actions/Twitter.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public func tweet(showComposeSheet: Bool = true) -> Action {
+    return Action(
+        identifier: "is.workflow.actions.tweet",
+        parameters: [
+            "WFSocialActionShowComposeSheet": showComposeSheet,
+            ]
+    )
+}
